Moved utility code living in the global Sandra namespace to Sandra.csproj.
