+++ Sandra.Chess/SpecializedArrays.cs
@@ -17,106 +17,6 @@
  * 
  *********************************************************************************/
 using System;
-using System.Linq;
-
-namespace Sandra
-{
-    /// <summary>
-    /// Contains an array which is indexed by an enumeration.
-    /// Always initialize such an array with the <see cref="New"/> method, e.g.:
-    /// <code>
-    /// var array = EnumIndexedArray&lt;StringComparison, StringComparer&gt;.New();
-    /// </code>
-    /// </summary>
-    /// <remarks>
-    /// Declaring an <see cref="EnumIndexedArray{TEnum, TValue}"/> with a non-enumeration key type
-    /// results in a <see cref="TypeInitializationException"/> being thrown.
-    /// </remarks>
-    public struct EnumIndexedArray<TEnum, TValue> where TEnum : struct
-    {
-        static EnumIndexedArray()
-        {
-            // Examine the enumeration.
-            TEnum[] values = EnumHelper<TEnum>.AllValues.ToArray();
-            for (int i = values.Length - 1; i >= 0; --i)
-            {
-                if ((int)(object)values[i] != i)
-                {
-                    throw new NotSupportedException("EnumIndexedArray<TEnum, TValue> does not support discontinuous enumerations, or enumerations that have a non-zero lower bound.");
-                }
-            }
-        }
-
-        private TValue[] arr;
-
-        private void init()
-        {
-            if (arr == null) arr = new TValue[Length];
-        }
-
-        /// <summary>
-        /// Initializes an empty array with default values.
-        /// </summary>
-        public static EnumIndexedArray<TEnum, TValue> New()
-        {
-            EnumIndexedArray<TEnum, TValue> wrapped = default(EnumIndexedArray<TEnum, TValue>);
-            wrapped.init();
-            return wrapped;
-        }
-
-        /// <summary>
-        /// Gets the number of keys in the array, which is equal to the number of members in the enumeration.
-        /// </summary>
-        public int Length => EnumHelper<TEnum>.EnumCount;
-
-        public TValue this[TEnum index]
-        {
-            get
-            {
-                try
-                {
-                    return arr[(int)(object)index];
-                }
-                catch (NullReferenceException)
-                {
-                    init();
-                    return arr[(int)(object)index];
-                }
-            }
-            set
-            {
-                try
-                {
-                    arr[(int)(object)index] = value;
-                }
-                catch (NullReferenceException)
-                {
-                    init();
-                    arr[(int)(object)index] = value;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Creates a copy of this array and returns it.
-        /// </summary>
-        public EnumIndexedArray<TEnum, TValue> Copy()
-        {
-            var copy = New();
-            try
-            {
-                Array.Copy(arr, copy.arr, Length);
-                return copy;
-            }
-            catch (NullReferenceException)
-            {
-                init();
-                Array.Copy(arr, copy.arr, Length);
-                return copy;
-            }
-        }
-    }
-}
 
 namespace Sandra.Chess
 {
